refactor(session): make Schedule to be lazy loaded
